+++ Editor/ValidationSuite/ValidationTests/ManifestValidation.cs
@@ -15,7 +15,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
         private const string UnityReleaseRegex = @"^[0-9]+[a|b|f]{1}[0-9]+$";
         internal static readonly int MinDescriptionSize = 50;
         internal static readonly int MaxDisplayNameLength = 50;
-        internal static readonly string docsFilePath = "manifest_validation_error.html";
+        internal static readonly string k_DocsFilePath = "manifest_validation_error.html";
 
         public ManifestValidation()
         {
@@ -50,7 +50,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             // Check package version, make sure it's a valid SemVer string.
             if (!SemVersion.TryParse(manifestData.version, out version))
             {
-                AddError("In package.json, \"version\" needs to be a valid \"Semver\". {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "version-needs-to-be-a-valid-semver"));
+                AddError("In package.json, \"version\" needs to be a valid \"Semver\". {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "version-needs-to-be-a-valid-semver"));
             }
         }
 
@@ -66,7 +66,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
                 // Check if the dependency semver is valid before doing anything else
                 SemVersion depVersion;
                 if (!SemVersion.TryParse(dependency.Value, out depVersion)) {
-                    AddError(@"In package.json, dependency ""{0}"" : ""{1}"" needs to be a valid ""Semver"". {2}", dependency.Key, dependency.Value, ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "dependency_needs_to_be_a_valid_Semver"));
+                    AddError(@"In package.json, dependency ""{0}"" : ""{1}"" needs to be a valid ""Semver"". {2}", dependency.Key, dependency.Value, ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "dependency_needs_to_be_a_valid_Semver"));
                     continue;
                 }
 
@@ -80,13 +80,16 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
                 
                 // Check if this package's dependencies are in production. That is a requirement for promotion.
 				// make sure to check the version actually resolved by upm and not the one necessarily listed by the package
+                // If the packageId is included in Context.packageIdsForPromotion then we can skip this check, as the package 
+                // is expected to be promoted by another process
                 var version = dependencyInfo != null ? dependencyInfo.version : dependency.Value;
                 var packageId = Utilities.CreatePackageId(dependency.Key, version);
                 
-                if (Context.ValidationType != ValidationType.VerifiedSet && !Utilities.PackageExistsOnProduction(packageId))
+                if (Context.ValidationType != ValidationType.VerifiedSet && (Context.packageIdsForPromotion == null || Context.packageIdsForPromotion.Length < 1 || !Context.packageIdsForPromotion.Contains(packageId)) && !Utilities.PackageExistsOnProduction(packageId))
                 {
+                    // ignore if  package is part of the context already
                     if (Context.ValidationType == ValidationType.Promotion || Context.ValidationType == ValidationType.AssetStore)
-                        AddError("Package dependency {0} is not promoted in production. {1}", packageId, ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "package-dependency-[packageID]-is-not-published-in-production"));
+                        AddError("Package dependency {0} is not promoted in production. {1}", packageId, ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "package-dependency-[packageID]-is-not-published-in-production"));
                     else
                         AddWarning("Package dependency {0} must be promoted to production before this package is promoted to production. (Except for core packages)", packageId);
                 }
@@ -111,53 +114,53 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             // This should probably be executed only in internal development, CI and Promotion contexts
             if (!PackageNamePrefixList.Any(namePrefix => (manifestData.name.StartsWith(namePrefix) && manifestData.name.Length > namePrefix.Length)))
             {
-                AddError("In package.json, \"name\" needs to start with one of these approved company names: {0}. {1}", string.Join(", ", PackageNamePrefixList), ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "name-needs-to-start-with-one-of-these-approved-company-names"));
+                AddError("In package.json, \"name\" needs to start with one of these approved company names: {0}. {1}", string.Join(", ", PackageNamePrefixList), ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "name-needs-to-start-with-one-of-these-approved-company-names"));
             }
 
             // There cannot be any capital letters in package names.
             if (manifestData.name.ToLower(CultureInfo.InvariantCulture) != manifestData.name)
             {
-                AddError("In package.json, \"name\" cannot contain capital letters. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "name-cannot-contain-capital-letters"));
+                AddError("In package.json, \"name\" cannot contain capital letters. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "name-cannot-contain-capital-letters"));
             }
 
             // Check name against our regex.
             Match match = Regex.Match(manifestData.name, UpmRegex);
             if (!match.Success)
             {
-                AddError("In package.json, \"name\" is not a valid name. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "name-is-not-a-valid-name"));
+                AddError("In package.json, \"name\" is not a valid name. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "name-is-not-a-valid-name"));
             }
 
 			// Package name cannot end with .framework, .plugin or .bundle.
             String[] strings = { ".framework", ".bundle", ".plugin" };
 			foreach (var value in strings) {
  				if (manifestData.name.EndsWith(value)) {
-					AddError("In package.json, \"name\" cannot end with .plugin, .bundle or .framework. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "name-cannot-end-with"));
+					AddError("In package.json, \"name\" cannot end with .plugin, .bundle or .framework. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "name-cannot-end-with"));
 				}
 			} 
 
             if (string.IsNullOrEmpty(manifestData.displayName))
             {
-                AddError("In package.json, \"displayName\" must be set. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "displayName-must-be-set"));
+                AddError("In package.json, \"displayName\" must be set. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "displayName-must-be-set"));
             }
             else if (manifestData.displayName.Length > MaxDisplayNameLength)
             {
-                AddError("In package.json, \"displayName\" is too long. Max Length = {0}. Current Length = {1}. {2}", MaxDisplayNameLength, manifestData.displayName.Length, ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "displayName-is-too-long"));
+                AddError("In package.json, \"displayName\" is too long. Max Length = {0}. Current Length = {1}. {2}", MaxDisplayNameLength, manifestData.displayName.Length, ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "displayName-is-too-long"));
             }
             else if (!Regex.Match(manifestData.displayName, UpmDisplayRegex).Success)
             {
-                AddError("In package.json, \"displayName\" cannot have any special characters. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "displayName-cannot-have-any-special-characters"));
+                AddError("In package.json, \"displayName\" cannot have any special characters. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "displayName-cannot-have-any-special-characters"));
             }
 
             // Check Description, make sure it's there, and not too short.
             if (manifestData.description.Length < MinDescriptionSize)
             {
-                AddError("In package.json, \"description\" is too short. Minimum Length = {0}. Current Length = {1}. {2}", MinDescriptionSize, manifestData.description.Length, ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "description-is-too-short"));
+                AddError("In package.json, \"description\" is too short. Minimum Length = {0}. Current Length = {1}. {2}", MinDescriptionSize, manifestData.description.Length, ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "description-is-too-short"));
             }
             
             // check unity field, if it's there
             if (!string.IsNullOrEmpty(manifestData.unity) && (manifestData.unity.Length > 6 || !Regex.Match(manifestData.unity, UnityRegex).Success))
             {
-                AddError($"In package.json, \"unity\" is invalid. It should only be <MAJOR>.<MINOR> (e.g. 2018.4). Current unity = {manifestData.unity}. {ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "unity-is-invalid")}");
+                AddError($"In package.json, \"unity\" is invalid. It should only be <MAJOR>.<MINOR> (e.g. 2018.4). Current unity = {manifestData.unity}. {ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "unity-is-invalid")}");
             }
             
             // check unityRelease field, if it's there
@@ -167,14 +170,14 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
                 if (!Regex.Match(manifestData.unityRelease, UnityReleaseRegex).Success)
                 {
                     AddError(
-                        $"In package.json, \"unityRelease\" is invalid. Current unityRelease = {manifestData.unityRelease}. {ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath, "unityrelease-is-invalid")}");
+                        $"In package.json, \"unityRelease\" is invalid. Current unityRelease = {manifestData.unityRelease}. {ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath, "unityrelease-is-invalid")}");
                 }
                 
                 // it should be accompanied of a unity field
                 if (string.IsNullOrEmpty(manifestData.unity))
                 {
                     AddError(
-                        $"In package.json, \"unityRelease\" needs a \"unity\" field to be used. {ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath, "unityrelease-without-unity")}");
+                        $"In package.json, \"unityRelease\" needs a \"unity\" field to be used. {ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath, "unityrelease-without-unity")}");
                 }
             }
 
@@ -189,9 +192,9 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
                 // Check if `repository.url` and `repository.revision` exist and the content is valid
                 string value;
                 if (!manifestData.repository.TryGetValue("url", out value) || string.IsNullOrEmpty(value))
-                    AddError("In package.json for a published package, there must be a \"repository.url\" field. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "for_a_published_package_there_must_be_a_repository.url_field"));
+                    AddError("In package.json for a published package, there must be a \"repository.url\" field. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "for_a_published_package_there_must_be_a_repository.url_field"));
                 if (!manifestData.repository.TryGetValue("revision", out value) || string.IsNullOrEmpty(value))
-                    AddError("In package.json for a published package, there must be a \"repository.revision\" field. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,  "for_a_published_package_there_must_be_a_repository.revision_field"));
+                    AddError("In package.json for a published package, there must be a \"repository.revision\" field. {0}", ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,  "for_a_published_package_there_must_be_a_repository.revision_field"));
             }
             else
             {
@@ -217,7 +220,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             {
                 // make sure it is not present in order to have a unified presentation of the author for all of our packages
                 AddError("A Unity package must not have an author field. Please remove the field. {0}",
-                    ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath,
+                    ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath,
                         "a_unity_package_must_not_have_an_author_field"));
             }
         }
@@ -230,7 +233,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             {
                 AddError(
                 "The `author` field is mandatory. Please add an `author` field in your package.json file",
-                ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath, "author_is_mandatory"));
+                ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath, "author_is_mandatory"));
                 return;
             }
 
@@ -244,7 +247,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             {
                 AddError(
                     "Invalid `author` field. The `author` field in your package.json file can be a string or an object ( name, email, url ), where `name` is mandatory. {0}",
-                    ErrorDocumentation.GetLinkMessage(ManifestValidation.docsFilePath, "author_is_invalid"));
+                    ErrorDocumentation.GetLinkMessage(ManifestValidation.k_DocsFilePath, "author_is_invalid"));
             }
         }
     }