+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.19.0] - 2020-12-15
+- Added support for excepting Asset Validation errors.
+- Changed the promoted dependencies check (`ManifestValidation`) to allow un-promoted dependencies if they're included in a list of packageIds provided to the Package Validation Suite
+- Fixed `System.NullReferenceException` during `API Validation`.
+
 ## [0.18.1] - 2020-11-17
 - Fixed `Editor` namespace clashing with `Editor` type (introduced in 0.18.0).